\name{extractoR.data-package}
\alias{extractoR.data-package}
\alias{extractoR.data}
\docType{package}
\encoding{UTF-8}
\title{R packages metadata extractor (data)}
\description{
  This package is part of extractoR which is a set of R packages that
  can be used to fetch, extract and query R package metadata.
  extractoR.data is used to read, manipulate and export R packages
  metadata.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\examples{
\dontrun{
}
}